# docs/swagger.yaml
basePath: /api
definitions:
  models.Board:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateListRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.List:
    properties:
      board:
        $ref: '#/definitions/models.Board'
      board_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      position:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.UpdateListRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
      position:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Kanban application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kanban API Documentation
  version: "1.0"
paths:
...
  /organizations/{orgID}/projects/{projectID}/boards/{boardID}/lists:
    get:
      description: Retrieves all lists within a specified board. User must own the
        board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of lists
          schema:
            items:
              $ref: '#/definitions/models.List'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all lists in a board
      tags:
      - Lists
    post:
      consumes:
      - application/json
      description: Creates a new list within a specified board. User must own the
        board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List creation details
        in: body
        name: list
        required: true
        schema:
          $ref: '#/definitions/models.CreateListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: List created successfully
          schema:
            $ref: '#/definitions/models.List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new list
      tags:
      - Lists
  /organizations/{orgID}/projects/{projectID}/boards/{boardID}/lists/{listID}:
    delete:
      description: Deletes a specific list by its ID within a specified board. User
        must own the board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a list
      tags:
      - Lists
    get:
      description: Retrieves a specific list by its ID within a specified board. User
        must own the board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List details
          schema:
            $ref: '#/definitions/models.List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list by ID
      tags:
      - Lists
    put:
      consumes:
      - application/json
      description: Updates a specific list by its ID within a specified board. User
        must own the board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      - description: List update details
        in: body
        name: list
        required: true
        schema:
          $ref: '#/definitions/models.UpdateListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List updated successfully
          schema:
            $ref: '#/definitions/models.List'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a list
      tags:
      - Lists
...
securityDefinitions:
  ApiKeyAuth:
    description: Enter your JWT token in the format "Bearer <your_token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...